Correct container limits in network container doc
Currently, the network container docs state that a project can have a maximum
of one container for each cloud provider. The page then goes on to list the
maximum number of Network Peering containers per cloud provider.

This commit tries to reconcile this information, while also adjusting
so that the list displays correctly.