Add stronger warning against attempting a shared tier upgrade (#516)
Users who attempt to upgrade from a shared tier cluster (M2/M5) to a dedicated tier cluster - even though it is called out as unsupported - will destroy the shared tier and create a dedicated tier.  This is due to the fact that upgrades are not supported yet Terraform will still try to make the changes it sees in the config.   Adding stronger language to the call out that this is not supported since it will result in data lost from the shared tier cluster.