fix: Interrupted creation of new cluster
When new cluster is created and interrupted during the creation process, the `SetId` function (set resource identifier) will not be called.
It is only called when the cluster is successfully created. Re-running tf apply will fail (cluster exist with same name).

This PR will call `SetId` directly after the create call to API whatever cluster is created successfully or not.
This will ensure that the cluster resource is imported in case of interrupted creation,
which is beneficial in case of automated CI/CD, pipelines, etc (no need to import resource manually in case of the cluster already exist).

Further, to achieve this we get rid of the cluster ID encoded property as `project_id`, `cluster_name` and `provider_name` are fairly enough
to uniquely identify the cluster resource.

Good example of similar behavior for example when create EKS cluster using
[AWS tf provider](https://github.com/hashicorp/terraform-provider-aws/blob/main/internal/service/eks/cluster.go#L360),
where cluster name is enough to identify the cluster and the `waitClusterCreated` is called after `SetId` which allow better handling for interrupted creation of resources.